merge slices of complex types
this is a breaking change in that it modifies the behavior of merging
when it comes to anything that cannot be compared explicitly by id.
in the past, with a few exceptions for primitive types, we would usually
just append the two slices together when merging, which is usually not
what you want. now, in the absence of an explicit id, we will hash the
entire item (which can be a complex struct) and store the hash in a lookup
table so that we know we've already seen it.

to handle the case when there are duplicates (b/c the primitive or complex
struct had equal values multiple times in the slice), we keep a counter
of how many we have seen. the merged result should equal the maximum number
of occurrences between the from and to sets.